refactor(renderer): split MarkdownRenderer into RenderConfig + Walker
Split MarkdownRenderer<B> into three types to fix a latent per-render
state-reset bug:

- RenderConfig (non-generic, settings only) in a new config.rs.
- MarkdownRenderer<B> (public facade) holding RenderConfig +
  processors + directives + PhantomData<B>.
- Walker<'r, B> (per-render scratch) in a new walker.rs, constructed
  fresh inside every render call and dropped on return.

## Why

Pre-refactor, MarkdownRenderer held per-render scratch (output
buffer, code-block index, heading id-counts, "seen first H1" flag,
scope stacks, text buffer, alert stack, list stack) alongside its
settings. The scratch was never reset between render calls on the
same &mut self, so reusing a renderer produced silently-wrong
output: heading IDs got stale suffixes, Confluence-mode title
extraction broke on the second render, code-block-processor
placeholders collided. All three downstream callers (rw-site,
rw-confluence, rw-kroki) worked around this by constructing a fresh
renderer per render. The type now enforces that.

## What

- New files: config.rs (settings struct + read-only helpers +
  WikilinkResolution + TitleResolver trait + 14 unit tests) and
  walker.rs (per-render scratch + 13 event-handling methods + 4
  unit tests). renderer.rs shrinks substantially on the non-test
  side, gains 5 regression tests.
- Public API surface unchanged. pub use renderer::{MarkdownRenderer,
  RenderResult} and pub use config::TitleResolver in lib.rs preserve
  all downstream usage; rw-site, rw-confluence, and rw-kroki compile
  and pass tests without source changes.
- Two load-bearing borrow patterns documented in walker.rs:
  - Pattern A: field-disjoint borrows on processors + output in the
    TagEnd::CodeBlock arm (NLL splits per-field; wrapping push_str
    in a &mut self method would break it).
  - Pattern B: tightly-scoped reborrow of directives in flush_text
    that releases before any &mut self method call. directives is
    Option<&'r mut DirectiveProcessor>, so access uses
    .as_deref_mut() rather than .as_mut().
- Free side benefit: panics inside a processor unwind through
  Walker, which is dropped on the stack; the facade's RenderConfig
  stays clean for the next render.
- debug_assert!s in Walker::finish catch malformed event streams
  (scopes/list_stack/alert_stack non-empty at end-of-render) in
  debug builds at zero release cost.

## Tests

327 unit tests pass (5 new regression tests covering id-counts,
Confluence-mode title, code-block-index, panic recovery, and
wikilink reuse, plus 14 config.rs and 4 walker.rs unit tests for
the helpers and per-render contracts), 29 doc tests, all downstream
test suites green. cargo clippy -p rw-renderer -- -D warnings clean,
0 cargo doc warnings.
